chore(rust): rename `ecma_script.rs` to `ecmascript.rs`
