+++ crates/rolldown/src/module_finalizers/isolating/impl_visit_mut.rs
@@ -4,7 +4,7 @@ use oxc::ast::VisitMut;
 use oxc::span::{CompactStr, Span, SPAN};
 use rolldown_common::{Interop, Module};
 use rolldown_ecmascript::TakeIn;
-use rolldown_utils::ecma_script::legitimize_identifier_name;
+use rolldown_utils::ecmascript::legitimize_identifier_name;
 
 use crate::utils::call_expression_ext::CallExpressionExt;
