+++ crates/rolldown_utils/src/ecmascript.rs
